refactor(metal): move RoPE out of attn.rs into rope.rs, no behaviour change
attn.rs 8998 -> 7999 lines. The types, the four kernels, both
encoders, the two host-upload launchers and their parity tests move
together; attn re-exports them at the paths ferrox-models reads. The
(source, entry point) pair for a pairing convention was restated in
the prefill pipeline warm-up beside the encoder's own match; it is one
table now, rope_kernel / rope_batch_kernel.
